style(arm support): rename arm interface file
